move Configuration.lintableFilesForPath out of extension and into framework
